chore: Introduce RocksMetaStore via RocksStore (unify functionality)
